+++ tests/fixtures/__init__.py
@@ -12,7 +12,7 @@ def patharg(path):
     return path.replace('\\', '\\\\\\')
 
 
-FIXTURES_ROOT = path.join(path.abspath(path.dirname(__file__)), 'fixtures')
+FIXTURES_ROOT = path.join(path.abspath(path.dirname(__file__)))
 FILE_PATH = path.join(FIXTURES_ROOT, 'test.txt')
 JSON_FILE_PATH = path.join(FIXTURES_ROOT, 'test.json')
 BIN_FILE_PATH = path.join(FIXTURES_ROOT, 'test.bin')