+++ CHANGELOG.rst
@@ -9,8 +9,11 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 -------------------------
 
 * Added ``Content-Type`` of files uploaded in ``multipart/form-data`` requests
+* Added ``--show-redirects, -R`` to show intermediate responses with ``--follow``
+* Added ``--max-redirects`` (default 30)
 * Added ``-A`` as short name for ``--auth-type``
-* Changed the default color style back to ``solarized`` as it supports
+* Added ``-F`` as short name for ``--follow``
+* Changed the default color style back to``solarized`` as it supports
   both the light and dark terminal background mode
 * Fixed ``--session`` when used with ``--download``
 * Fixed handling of ``Content-Type`` with multiple ``+subtype`` parts